requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
icoextract==0.1.5
idna==3.7
pefile==2023.2.7
pillow==10.3.0
psutil==5.9.8
python-dotenv==1.0.1
requests==2.32.3
ttkbootstrap==1.10.1
urllib3==2.2.1
